setwait no longer caps at 60fps, so "frameskip" debug key can increase framerate 5x properly
(Recall it doesn't actually do any frame skipping, and it's probably not
going to work with vsync in gfx_directx.)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7230 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ game.bas
@@ -667,7 +667,7 @@ gam.walk_through_walls = NO
 'DEBUG debug "pre-call update_heroes"
 update_heroes(YES)
 setkeys
-DIM speedcontrol_this_tick as integer = speedcontrol
+DIM speedcontrol_this_tick as double = speedcontrol
 DO
  'DEBUG debug "top of master loop"
  setwait speedcontrol_this_tick
